Fix domain node not visited when chapel frontend version 1.25+ is used
- Between some versions of chapel while they were trying
their new frontend compiler next, they  restricted domains
and array parsing in some way.

- Previously var: data [{1..5}] real; was well defined
however, var: data [1..5] real; was allowed or to me at least,
it is a hint that it was discouraged.

- When I tried our old implementation, i see that we enter the
domain node before we enter a range node. With frontend after
release 1.25.0 instead of domain node for 1..5 we enter
directly into the range node.

- Ref: https://github.com/chapel-lang/chapel/pull/19187

Signed-off-by: Shreyas Atre <shreyasatre16@gmail.com>